:recycle: QD-14839: gate Linux-only linter tests with //go:build linux, drop needs.GOOS
