+++ cdnet/cdnet_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 /*
  * Copyright 2021-2024 JetBrains s.r.o.
  *
