+++ clang/run_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package main
 
 import (
